[ENG-1095] HEIF* and PDF thumbnail orientation (#1815)
* fix horizontal pdf rendering

* don't rotate certain formats (heif*)

* return `u32` from scaling function

* add `avifs` as a valid extension

* add avifs as a valid MIME type

* re-categorize `avifs` as a video (sequence)

* Fix mime-types

* Add Hif extension
 - Cannon uses a special extension for heif in its cameras: https://github.com/digipres/digipres.github.io/blob/master/_sources/registries/fdd/fddXML/fdd000525.xml#L213-L221

* hif is actually a heic

---------

+++ crates/images/src/svg.rs
@@ -31,9 +31,9 @@ impl ImageHandler for SvgHandler {
 			scale_dimensions(rtree.size.width(), rtree.size.height(), SVG_TARGET_PX);
 
 		let size = if rtree.size.width() > rtree.size.height() {
-			rtree.size.to_int_size().scale_to_width(scaled_w as u32)
+			rtree.size.to_int_size().scale_to_width(scaled_w)
 		} else {
-			rtree.size.to_int_size().scale_to_height(scaled_h as u32)
+			rtree.size.to_int_size().scale_to_height(scaled_h)
 		}
 		.ok_or(Error::InvalidLength)?;
